Refactor: HomeHeader 공용 컴포넌트로 이동 (#18)
